474. Ones and Zeroes
